fileutil: move LUtimesNano to fileutil
